+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Ember Changelog
 
-## v6.9.0-beta.1 (October 13, 2025)
+## v6.10.0-beta.1 (November 25, 2025)
+
+- [#21013](https://github.com/emberjs/ember.js/pull/21013) / [#21009](https://github.com/emberjs/ember.js/pull/21009) [DEPRECATION] Deprecate use of amd bundles per [RFC #1101](https://rfcs.emberjs.com/id/1101-deprecate-ember-vendor-bundles).
+- [#21004](https://github.com/emberjs/ember.js/pull/21004) Remove deprecated TypeScript compiler options for TS 6.0 compatibility
+
+## v6.9.0 (November 25, 2025)
 
 - All changes were internal, docs, and/or bugfixes that were backported.
+
+## v6.8.2 (November 17, 2025)
+
+- [#21002](https://github.com/emberjs/ember.js/pull/21002) [BUGFIX] Don't render if we're mid-destroy
+
+## v6.8.1 (October 30, 2025)
+
+- [#20994](https://github.com/emberjs/ember.js/pull/20994) [BUGFIX] Fix loose-mode components invoked from a strict mode component via renderComponent
 
 ## v6.8.0 (October 13, 2025)